use double

+++ src/main/kotlin/eu/aagsolutions/trip/service/geo/model/GeoPoint.kt
@@ -1,5 +1,3 @@
 package eu.aagsolutions.trip.service.geo.model
 
-import java.math.BigDecimal
-
-class GeoPoint(val address: String, val lat: BigDecimal?, val lon: BigDecimal?)
+class GeoPoint(val address: String, val lat: Double?, val lon: Double?)
